Create Network Scanner Tool
I use VM's regularly. My first step after start up is to run updates and check health. So, I wanted to automate this process and share it. This was done in Bash obviously and the tools needed are Nmap, Logcheck, and OpenVas. You can do this in Windows using Bash for Windows or setting up Bash environment for windows. 
Important to note that you must input your target ip and the info for your OpenVas.